+++ graph.cpp
@@ -7,16 +7,16 @@ void prnt_grph();
 ***/
 void graph_init()
 {
-    graph = (vnode_t<size_t>**)malloc(sizeof(vnode_t<size_t>*)*num_prnts);
+    graph = (vnode_t<uint32_t>**)malloc(sizeof(vnode_t<uint32_t>*)*num_prnts);
 
     // Initialize empty graph
-    for (size_t i = 0; i < num_prnts; i++)
+    for (uint32_t i = 0; i < num_prnts; i++)
     {
         graph[i] = NULL;
     }
 
     // Add edges given the fields
-    for (size_t i = 0; i < num_prnts; i++)
+    for (uint32_t i = 0; i < num_prnts; i++)
     {
         vnode_t<field_t>* lvnode = prnts[i].fields;
 
@@ -24,13 +24,13 @@ void graph_init()
         while (lvnode)
         {
             field_t& lfield = lvnode->val;
-            size_t lprim_sz = lfield.prim_sz;
+            uint64_t lprim_sz = lfield.prim_sz;
 
             // Add edge for a pointer
             if (!lprim_sz)
             {
-                size_t ltype = lfield.type;
-                graph[ltype] = vnode_distinct_insert<size_t>(graph[ltype], i);
+                uint64_t ltype = lfield.type;
+                graph[ltype] = vnode_distinct_insert<uint32_t>(graph[ltype], i);
             }
 
             lvnode = lvnode->next;
@@ -45,21 +45,21 @@ void graph_init()
 ***/
 void prnt_grph()
 {
-    for (size_t i = 0; i < num_prnts; i++)
+    for (uint32_t i = 0; i < num_prnts; i++)
     {
-        fprintf(stderr, "%ld:", i);
+        fprintf(stderr, "%u:", i);
 
-        vnode_t<size_t>* lvnode = graph[i];
+        vnode_t<uint32_t>* lvnode = graph[i];
 
         while (lvnode && (lvnode->next))
         {
-            fprintf(stderr, "%ld->", lvnode->val);
+            fprintf(stderr, "%u->", lvnode->val);
             lvnode = lvnode->next;
         }
 
         if (lvnode)
         {
-            fprintf(stderr, "%ld", lvnode->val);
+            fprintf(stderr, "%u", lvnode->val);
         }
 
         fprintf(stderr, "\n");
